Move main function to a tests folder, and convert the data structure executables to libraries. Link to tests and new included headers
