+++ CMakeLists.txt
@@ -11,4 +11,6 @@ else()
     message("RELEASE VERSION")
 endif()
 
+include_directories(include)
 add_subdirectory(src)
+add_subdirectory(tests)
