+++ include/fibonacci_arr.h
@@ -1,0 +1,10 @@
+/*
+ * Fibonacci via Array Cache
+ * By: Liam P. Walsh
+*/
+
+#include <stdlib.h>
+#include <stdio.h>
+#define ARRAY_SIZE 45
+
+unsigned int fibonacci(unsigned int i, unsigned int array[]);
