+++ include/fibonacci_loop.h
@@ -1,0 +1,12 @@
+/*
+ * Fibonacci via loop
+ * By: Liam P. Walsh
+ * Uses a For Loop with the formula as variables.
+*/
+
+#include <stdlib.h>
+#include <stdio.h>
+
+// Uses three variables to iterate with instead of recursion.
+// Follows the formula: Fib(n) = Fib(n-2) + Fib(n-1)
+unsigned int fibonacci(unsigned int i);
