+++ include/fibonacci_rec.h
@@ -1,0 +1,9 @@
+/*
+ * Fibonacci via Recursive Function.
+ * By: Liam P. Walsh
+*/
+
+#include <stdlib.h>
+#include <stdio.h>
+
+unsigned int fibonacci(unsigned int i);
