+++ src/fibonacci_arr.c
@@ -5,7 +5,7 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-#define ARRAY_SIZE 45
+#include "fibonacci_arr.h"
 
 unsigned int fibonacci(unsigned int i, unsigned int array[]) {
   // Determine if number is too big for array.
@@ -24,16 +24,4 @@ unsigned int fibonacci(unsigned int i, unsigned int array[]) {
     array[i] = fibonacci(i-1, array) + fibonacci(i-2, array);
   }
   return array[i];
-}
-
-int main(void)
-{
-  // Array cache
-  unsigned int array[ARRAY_SIZE] = {0};
-  printf("Fib %u: %u\n", 5, fibonacci(5, array));
-  printf("Fib %u: %u\n", 7, fibonacci(7, array));
-  printf("Fib %u: %u\n", 5, fibonacci(5, array));
-  printf("Fib %u: %u\n", 43, fibonacci(43, array));
-  printf("Fib %u: %u\n", 40, fibonacci(44, array));
-  return 0;
 }
