+++ src/fibonacci_loop.c
@@ -6,6 +6,7 @@
 
 #include <stdlib.h>
 #include <stdio.h>
+#include "fibonacci_loop.h"
 
 // Uses three variables to iterate with instead of recursion.
 // Follows the formula: Fib(n) = Fib(n-2) + Fib(n-1)
@@ -31,14 +32,4 @@ unsigned int fibonacci(unsigned int i) {
 
   // Return sum unless given value was 0.
   return i > 0 ? sum : n_minus_2;
-}
-
-int main(int argc, char *argv[]) {
-  printf("Fib %u: %u\n", 1, fibonacci(0));
-  printf("Fib %u: %u\n", 1, fibonacci(1));
-  printf("Fib %u: %u\n", 2, fibonacci(2));
-  printf("Fib %u: %u\n", 5, fibonacci(5));
-  printf("Fib %u: %u\n", 15, fibonacci(15));
-  printf("Fib %u: %u\n", 9, fibonacci(9));
-  return 0;
 }
