+++ src/single_linked_list.c
@@ -5,20 +5,7 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-
-// Declares linked node struct.
-typedef struct linked_node_ {
-  // Data of linked list node
-  int number;
-  // The next node
-  struct linked_node_ * node;
-} linked_node;
-
-// Declares linked list struct. (hold first node)
-typedef struct linked_list_ {
-  // The first node of the list
-  linked_node * head;
-} linked_list;
+#include "single_linked_list.h"
 
 // Allocates node with the given data, and returns address.
 linked_node * create_node(int number) {
@@ -39,39 +26,39 @@ linked_list * create_list(linked_node * head) {
 }
 
 // Append node to front of list. (Replaces head node)
-int append_front_node(linked_node * new, linked_list * list) {
+int append_front_node(linked_node * new_node, linked_list * list) {
 	// Makes sure there is a list and node to append.
-	if (!list || !new) {
+	if (!list || !new_node) {
 		return 0;
 	}
 	// If the list is empty, sets the node as head.
 	if (!list->head) {
-		list->head = new;
+		list->head = new_node;
 	}
 	// Replaces the current head of the list.
 	else {
-		new->node = list->head;
-		list->head = new;
+		new_node->node = list->head;
+		list->head = new_node;
 	}
 	return 1;
 }
 
 // Append node to end of list. (Parses for tail)
-int append_back_node(linked_node * new, linked_list * list) {
+int append_back_node(linked_node * new_node, linked_list * list) {
 	// Makes sure there is a list and node to append.
-	if (!list || !new) {
+	if (!list || !new_node) {
 		return 0;
 	}
 	// If the list is empty, sets the node as head.
 	if (!list->head) {
-		list->head = new;
+		list->head = new_node;
 	}
 	// Finds the end of the list, and appends to the end.
 	else {
 		linked_node *node = list->head;
 		while (node) {
 			if (!node->node) {
-				node->node = new;
+				node->node = new_node;
 				break;
 			}
 			node = node->node;
@@ -143,42 +130,4 @@ void print_list(linked_list * list) {
 		printf("   Current Node's int = %d\n", node->number);
 		node = node->node;
 	}
-}
-
-int main()
-{
-  printf("Creating Linked list:\n");
-  linked_list * list = calloc(1, sizeof(linked_list));
-	list->head = create_node(3);
-  print_list(list);
-	if (!append_back_node(create_node(1), list)) {
-		printf("Append Back failed.\n");
-	}
-  print_list(list);
-	if (!append_back_node(create_node(2), list)) {
-		printf("Append Back failed.\n");
-	}
-  print_list(list);
-	if (!append_back_node(create_node(3), list)) {
-		printf("Append Back failed.\n");
-	}
-  print_list(list);
-	if (!append_front_node(create_node(1), list)) {
-		printf("Append Front failed.\n");
-	}
-  print_list(list);
-
-	printf("Removing nodes with value 1:\n");
-	if (!search_and_destroy_node(list, 1)) {
-		printf("Search and Destroy failed.\n");
-	}
-	print_list(list);
-
-	printf("Deleting Linked list:\n");
-	if (!empty_list(list)) {
-		printf("Empty List failed.\n");
-	}
-	print_list(list);
-	free(list);
-  return 0;
 }
